Chunking is working

+++ src/tileset.h
@@ -9,4 +9,4 @@ typedef struct
 } tileset_t;
 
 void tileset_create(const char* path, int tile_width);
-void tileset_draw_id(int id, int x, int y);
+void tileset_draw_id(int id, float x, float y);
